# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/create-category": {
            "post": {
                "tags": [
                    "Category Master"
                ],
                "summary": "Create Category",
                "description": "Create Category",
                "operationId": "createCategory",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter Category Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Category 1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/edit-category": {
            "post": {
                "tags": [
                    "Category Master"
                ],
                "summary": "Edit Category",
                "description": "Edit Category",
                "operationId": "editCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Category Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter Category Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Home Appliances"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category-list": {
            "post": {
                "tags": [
                    "Category Master"
                ],
                "summary": "Category List",
                "description": "Category List",
                "operationId": "categoryList",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "no of record you already get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "0"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "no of record you want to get",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by category name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Category Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Category Status (Active or Inactive)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column (name)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/view-category": {
            "post": {
                "tags": [
                    "Category Master"
                ],
                "summary": "View Category",
                "description": "View Category",
                "operationId": "viewCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Category Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category-status-change": {
            "post": {
                "tags": [
                    "Category Master"
                ],
                "summary": "Category Status Change",
                "description": "Category Status Change",
                "operationId": "categoryStatusChange",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Category Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Enter Status (Active or Inactive)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Inactive"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category-delete": {
            "post": {
                "tags": [
                    "Category Master"
                ],
                "summary": "Category Delete",
                "description": "Category Delete",
                "operationId": "categoryDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Category Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}